Move graph perm importance to correct file
